[v0.86][tools] Complete structural cleanup of oversized PR control-plane and artifact surfaces (Closes #1283)
