+++ adl/src/cli/run_artifacts/cognitive.rs
@@ -1,626 +1,4 @@
-use anyhow::{anyhow, Context, Result};
-use serde::{Deserialize, Serialize};
-use std::collections::{BTreeMap, BTreeSet, HashMap, HashSet};
-use std::path::{Path, PathBuf};
-
-use ::adl::{artifacts, execute, failure_taxonomy, instrumentation, resolve, trace};
-
-pub(crate) use super::run_artifacts_types::*;
-
-pub(crate) fn stable_fingerprint_hex(bytes: &[u8]) -> String {
-    let mut hash: u64 = 0xcbf29ce484222325;
-    for b in bytes {
-        hash ^= u64::from(*b);
-        hash = hash.wrapping_mul(0x100000001b3);
-    }
-    format!("{hash:016x}")
-}
-
-pub(crate) fn execution_plan_hash<T: Serialize>(plan: &T) -> Result<String> {
-    let plan_json = serde_json::to_vec(plan).context("serialize execution plan for hashing")?;
-    Ok(stable_fingerprint_hex(&plan_json))
-}
-
-pub(crate) fn classify_failure_kind(err: &anyhow::Error) -> Option<&'static str> {
-    failure_taxonomy::classify(err)
-}
-
-pub(crate) fn build_run_summary(
-    resolved: &resolve::AdlResolved,
-    status: &str,
-    pause: Option<&execute::PauseState>,
-    steps: &[StepStateArtifact],
-    records: usize,
-    failure: Option<&anyhow::Error>,
-    run_paths: &artifacts::RunArtifactPaths,
-) -> RunSummaryArtifact {
-    let failed_steps = steps.iter().filter(|s| s.status == "failure").count();
-    let completed_steps = steps
-        .iter()
-        .filter(|s| s.status == "success" || s.status == "failure")
-        .count();
-    let delegation_steps = resolved
-        .steps
-        .iter()
-        .filter(|s| {
-            s.delegation
-                .as_ref()
-                .map(|d| !d.is_effectively_empty())
-                .unwrap_or(false)
-        })
-        .count();
-    let delegation_requires_verification_steps = resolved
-        .steps
-        .iter()
-        .filter(|s| {
-            s.delegation
-                .as_ref()
-                .and_then(|d| d.requires_verification)
-                .unwrap_or(false)
-        })
-        .count();
-    let mut security_denials_by_code = BTreeMap::new();
-    if let Some(code) = failure.and_then(classify_failure_kind) {
-        *security_denials_by_code
-            .entry(code.to_string())
-            .or_insert(0) += 1;
-    }
-
-    let (
-        security_envelope_enabled,
-        signing_required,
-        key_id_required,
-        mut allowed_algs,
-        mut allowed_key_sources,
-    ) = if let Some(remote) = resolved.doc.run.remote.as_ref() {
-        (
-            true,
-            remote.require_signed_requests,
-            remote.require_key_id,
-            remote.verify_allowed_algs.clone(),
-            remote.verify_allowed_key_sources.clone(),
-        )
-    } else {
-        (false, false, false, Vec::new(), Vec::new())
-    };
-    allowed_algs.sort();
-    allowed_algs.dedup();
-    allowed_key_sources.sort();
-    allowed_key_sources.dedup();
-    let scores_rel = run_paths
-        .scores_json()
-        .strip_prefix(run_paths.run_dir())
-        .map(|p| p.display().to_string())
-        .unwrap_or_else(|_| "learning/scores.json".to_string());
-    let suggestions_rel = run_paths
-        .suggestions_json()
-        .strip_prefix(run_paths.run_dir())
-        .map(|p| p.display().to_string())
-        .unwrap_or_else(|_| "learning/suggestions.json".to_string());
-    let aee_decision_rel = run_paths
-        .aee_decision_json()
-        .strip_prefix(run_paths.run_dir())
-        .map(|p| p.display().to_string())
-        .unwrap_or_else(|_| "learning/aee_decision.json".to_string());
-    let cognitive_signals_rel = run_paths
-        .cognitive_signals_json()
-        .strip_prefix(run_paths.run_dir())
-        .map(|p| p.display().to_string())
-        .unwrap_or_else(|_| "learning/cognitive_signals.v1.json".to_string());
-    let fast_slow_path_rel = run_paths
-        .fast_slow_path_json()
-        .strip_prefix(run_paths.run_dir())
-        .map(|p| p.display().to_string())
-        .unwrap_or_else(|_| "learning/fast_slow_path.v1.json".to_string());
-    let agency_selection_rel = run_paths
-        .agency_selection_json()
-        .strip_prefix(run_paths.run_dir())
-        .map(|p| p.display().to_string())
-        .unwrap_or_else(|_| "learning/agency_selection.v1.json".to_string());
-    let bounded_execution_rel = run_paths
-        .bounded_execution_json()
-        .strip_prefix(run_paths.run_dir())
-        .map(|p| p.display().to_string())
-        .unwrap_or_else(|_| "learning/bounded_execution.v1.json".to_string());
-    let evaluation_signals_rel = run_paths
-        .evaluation_signals_json()
-        .strip_prefix(run_paths.run_dir())
-        .map(|p| p.display().to_string())
-        .unwrap_or_else(|_| "learning/evaluation_signals.v1.json".to_string());
-    let cognitive_arbitration_rel = run_paths
-        .cognitive_arbitration_json()
-        .strip_prefix(run_paths.run_dir())
-        .map(|p| p.display().to_string())
-        .unwrap_or_else(|_| "learning/cognitive_arbitration.v1.json".to_string());
-    let affect_state_rel = run_paths
-        .affect_state_json()
-        .strip_prefix(run_paths.run_dir())
-        .map(|p| p.display().to_string())
-        .unwrap_or_else(|_| "learning/affect_state.v1.json".to_string());
-    let reasoning_graph_rel = run_paths
-        .reasoning_graph_json()
-        .strip_prefix(run_paths.run_dir())
-        .map(|p| p.display().to_string())
-        .unwrap_or_else(|_| "learning/reasoning_graph.v1.json".to_string());
-    let cluster_groundwork_rel = run_paths
-        .cluster_groundwork_json()
-        .strip_prefix(run_paths.run_dir())
-        .map(|p| p.display().to_string())
-        .unwrap_or_else(|_| "meta/cluster_groundwork.json".to_string());
-
-    RunSummaryArtifact {
-        run_summary_version: RUN_SUMMARY_VERSION,
-        artifact_model_version: artifacts::ARTIFACT_MODEL_VERSION,
-        run_id: resolved.run_id.clone(),
-        workflow_id: resolved.workflow_id.clone(),
-        adl_version: resolved.doc.version.clone(),
-        swarm_version: env!("CARGO_PKG_VERSION").to_string(),
-        status: status.to_string(),
-        error_kind: failure.and_then(classify_failure_kind).map(str::to_string),
-        counts: RunSummaryCounts {
-            total_steps: resolved.steps.len(),
-            completed_steps,
-            failed_steps,
-            provider_call_count: records,
-            delegation_steps,
-            delegation_requires_verification_steps,
-        },
-        policy: RunSummaryPolicy {
-            security_envelope_enabled,
-            signing_required,
-            key_id_required,
-            verify_allowed_algs: allowed_algs,
-            verify_allowed_key_sources: allowed_key_sources,
-            sandbox_policy: "centralized_path_resolver_v1".to_string(),
-            security_denials_by_code,
-        },
-        links: RunSummaryLinks {
-            run_json: "run.json".to_string(),
-            steps_json: "steps.json".to_string(),
-            pause_state_json: pause.map(|_| "pause_state.json".to_string()),
-            outputs_dir: run_paths
-                .outputs_dir()
-                .strip_prefix(run_paths.run_dir())
-                .map(|p| p.display().to_string())
-                .unwrap_or_else(|_| "outputs".to_string()),
-            logs_dir: run_paths
-                .logs_dir()
-                .strip_prefix(run_paths.run_dir())
-                .map(|p| p.display().to_string())
-                .unwrap_or_else(|_| "logs".to_string()),
-            learning_dir: run_paths
-                .learning_dir()
-                .strip_prefix(run_paths.run_dir())
-                .map(|p| p.display().to_string())
-                .unwrap_or_else(|_| "learning".to_string()),
-            scores_json: Some(scores_rel),
-            suggestions_json: Some(suggestions_rel),
-            aee_decision_json: Some(aee_decision_rel),
-            cognitive_signals_json: Some(cognitive_signals_rel),
-            fast_slow_path_json: Some(fast_slow_path_rel),
-            agency_selection_json: Some(agency_selection_rel),
-            bounded_execution_json: Some(bounded_execution_rel),
-            evaluation_signals_json: Some(evaluation_signals_rel),
-            cognitive_arbitration_json: Some(cognitive_arbitration_rel),
-            affect_state_json: Some(affect_state_rel),
-            reasoning_graph_json: Some(reasoning_graph_rel),
-            overlays_dir: run_paths
-                .overlays_dir()
-                .strip_prefix(run_paths.run_dir())
-                .map(|p| p.display().to_string())
-                .unwrap_or_else(|_| "learning/overlays".to_string()),
-            cluster_groundwork_json: run_paths
-                .cluster_groundwork_json()
-                .is_file()
-                .then_some(cluster_groundwork_rel),
-            trace_json: None,
-        },
-    }
-}
-
-pub(crate) fn build_cluster_groundwork_artifact(
-    resolved: &resolve::AdlResolved,
-    steps: &[StepStateArtifact],
-    tr: &trace::Trace,
-) -> ClusterGroundworkArtifact {
-    let mut remaining_deps: BTreeMap<String, BTreeSet<String>> = resolved
-        .execution_plan
-        .nodes
-        .iter()
-        .map(|node| {
-            (
-                node.step_id.clone(),
-                node.depends_on.iter().cloned().collect::<BTreeSet<_>>(),
-            )
-        })
-        .collect();
-    let mut remaining_nodes: BTreeSet<String> =
-        remaining_deps.keys().cloned().collect::<BTreeSet<_>>();
-    let mut readiness_frontiers = Vec::new();
-    while !remaining_nodes.is_empty() {
-        let ready = remaining_nodes
-            .iter()
-            .filter(|step_id| {
-                remaining_deps
-                    .get(step_id.as_str())
-                    .map(|deps| deps.is_empty())
-                    .unwrap_or(false)
-            })
-            .cloned()
-            .collect::<Vec<_>>();
-        if ready.is_empty() {
-            break;
-        }
-        readiness_frontiers.push(ClusterReadyFrontier {
-            frontier_index: readiness_frontiers.len() as u32,
-            ready_step_ids: ready.clone(),
-        });
-        for step_id in &ready {
-            remaining_nodes.remove(step_id);
-        }
-        for deps in remaining_deps.values_mut() {
-            for step_id in &ready {
-                deps.remove(step_id);
-            }
-        }
-    }
-
-    let mut attempts_by_step: BTreeMap<String, u32> = BTreeMap::new();
-    for event in &tr.events {
-        if let trace::TraceEvent::StepStarted { step_id, .. } = event {
-            *attempts_by_step.entry(step_id.clone()).or_insert(0) += 1;
-        }
-    }
-    let status_by_step = steps
-        .iter()
-        .map(|step| (step.step_id.clone(), step.status.clone()))
-        .collect::<BTreeMap<_, _>>();
-    let depends_on_by_step = resolved
-        .execution_plan
-        .nodes
-        .iter()
-        .map(|node| (node.step_id.clone(), node.depends_on.clone()))
-        .collect::<BTreeMap<_, _>>();
-
-    let mut lease_records = Vec::new();
-    let mut issued_sequence: u32 = 0;
-    for frontier in &readiness_frontiers {
-        for step_id in &frontier.ready_step_ids {
-            issued_sequence = issued_sequence.saturating_add(1);
-            let status = status_by_step
-                .get(step_id)
-                .map(|value| value.as_str())
-                .unwrap_or("not_run");
-            let lease_state = match status {
-                "success" => "completed",
-                "failure" => "failed",
-                _ => "planned",
-            };
-            lease_records.push(ClusterLeaseRecord {
-                issued_sequence,
-                lease_id: format!("lease:{}:{}:1", resolved.run_id, step_id),
-                step_id: step_id.clone(),
-                depends_on: depends_on_by_step.get(step_id).cloned().unwrap_or_default(),
-                observed_attempts: attempts_by_step.get(step_id).copied().unwrap_or(0),
-                claim_owner: "adl-coordinator-local".to_string(),
-                worker_id: "adl-worker-local".to_string(),
-                lease_state: lease_state.to_string(),
-            });
-        }
-    }
-
-    ClusterGroundworkArtifact {
-        cluster_groundwork_version: CLUSTER_GROUNDWORK_VERSION,
-        run_id: resolved.run_id.clone(),
-        workflow_id: resolved.workflow_id.clone(),
-        coordinator_id: "adl-coordinator-local".to_string(),
-        worker_id: "adl-worker-local".to_string(),
-        canonical_ordering_key: "(run_id, step_id, attempt)".to_string(),
-        frontier_ordering: "topological_frontier_then_step_id".to_string(),
-        readiness_frontiers,
-        lease_records,
-    }
-}
-
-pub(crate) fn build_run_status(
-    resolved: &resolve::AdlResolved,
-    tr: &trace::Trace,
-    overall_status: &str,
-    steps: &[StepStateArtifact],
-    failure: Option<&anyhow::Error>,
-    resume_completed_step_ids: &BTreeSet<String>,
-) -> RunStatusArtifact {
-    let mut completed_steps: BTreeSet<String> = resume_completed_step_ids.clone();
-    let mut pending_steps: BTreeSet<String> = BTreeSet::new();
-    let mut failed_step_id: Option<String> = None;
-
-    for step in steps {
-        match step.status.as_str() {
-            "success" => {
-                completed_steps.insert(step.step_id.clone());
-            }
-            "failure" => {
-                if failed_step_id.is_none() {
-                    failed_step_id = Some(step.step_id.clone());
-                }
-                pending_steps.insert(step.step_id.clone());
-            }
-            _ => {
-                pending_steps.insert(step.step_id.clone());
-            }
-        }
-    }
-
-    let mut attempts_by_step: BTreeMap<String, u32> = resume_completed_step_ids
-        .iter()
-        .map(|step_id| (step_id.clone(), 0))
-        .collect();
-    let mut started_set = BTreeSet::new();
-    for event in &tr.events {
-        if let trace::TraceEvent::StepStarted { step_id, .. } = event {
-            started_set.insert(step_id.clone());
-            *attempts_by_step.entry(step_id.clone()).or_insert(0) += 1;
-        }
-    }
-
-    let scheduler_policy = execute::scheduler_policy_for_run(resolved).ok().flatten();
-
-    RunStatusArtifact {
-        run_status_version: RUN_STATUS_VERSION,
-        run_id: resolved.run_id.clone(),
-        workflow_id: resolved.workflow_id.clone(),
-        overall_status: overall_status.to_string(),
-        failure_kind: failure.and_then(classify_failure_kind).map(str::to_string),
-        failed_step_id,
-        completed_steps: completed_steps.into_iter().collect(),
-        pending_steps: pending_steps.into_iter().collect(),
-        started_steps: if started_set.is_empty() {
-            None
-        } else {
-            Some(started_set.into_iter().collect())
-        },
-        attempt_counts_by_step: attempts_by_step,
-        effective_max_concurrency: scheduler_policy.map(|(value, _)| value),
-        effective_max_concurrency_source: scheduler_policy
-            .map(|(_, source)| source.as_str().to_string()),
-    }
-}
-
-pub(crate) fn compute_retry_count(tr: &trace::Trace) -> usize {
-    let mut started_by_step: BTreeMap<&str, usize> = BTreeMap::new();
-    for event in &tr.events {
-        if let trace::TraceEvent::StepStarted { step_id, .. } = event {
-            *started_by_step.entry(step_id.as_str()).or_insert(0) += 1;
-        }
-    }
-    started_by_step
-        .values()
-        .map(|count| count.saturating_sub(1))
-        .sum()
-}
-
-pub(crate) fn compute_max_parallel_observed(tr: &trace::Trace) -> usize {
-    let mut active: BTreeSet<&str> = BTreeSet::new();
-    let mut max_parallel = 0usize;
-    for event in &tr.events {
-        match event {
-            trace::TraceEvent::StepStarted { step_id, .. } => {
-                active.insert(step_id.as_str());
-                max_parallel = max_parallel.max(active.len());
-            }
-            trace::TraceEvent::StepFinished { step_id, .. } => {
-                active.remove(step_id.as_str());
-            }
-            _ => {}
-        }
-    }
-    max_parallel
-}
-
-pub(crate) fn build_scores_artifact(
-    run_summary: &RunSummaryArtifact,
-    tr: &trace::Trace,
-) -> ScoresArtifact {
-    let success_steps = run_summary
-        .counts
-        .completed_steps
-        .saturating_sub(run_summary.counts.failed_steps);
-    let success_ratio = if run_summary.counts.total_steps == 0 {
-        1.0
-    } else {
-        let permille = (success_steps * 1000) / run_summary.counts.total_steps;
-        (permille as f64) / 1000.0
-    };
-    let security_denied_count: usize = run_summary.policy.security_denials_by_code.values().sum();
-    let delegation_denied_count: usize = run_summary
-        .policy
-        .security_denials_by_code
-        .iter()
-        .filter_map(|(code, count)| {
-            if code.starts_with("DELEGATION_") {
-                Some(*count)
-            } else {
-                None
-            }
-        })
-        .sum();
-
-    ScoresArtifact {
-        scores_version: SCORES_VERSION,
-        run_id: run_summary.run_id.clone(),
-        generated_from: ScoresGeneratedFrom {
-            artifact_model_version: run_summary.artifact_model_version,
-            run_summary_version: run_summary.run_summary_version,
-        },
-        summary: ScoresSummary {
-            success_ratio,
-            failure_count: run_summary.counts.failed_steps,
-            retry_count: compute_retry_count(tr),
-            delegation_denied_count,
-            security_denied_count,
-        },
-        metrics: ScoresMetrics {
-            scheduler_max_parallel_observed: compute_max_parallel_observed(tr),
-        },
-    }
-}
-
-pub(crate) fn read_scores_if_present(
-    run_paths: &artifacts::RunArtifactPaths,
-) -> Option<ScoresArtifact> {
-    let path = run_paths.scores_json();
-    let raw = std::fs::read_to_string(path).ok()?;
-    serde_json::from_str::<ScoresArtifact>(&raw).ok()
-}
-
-pub(crate) fn build_suggestions_artifact(
-    run_summary: &RunSummaryArtifact,
-    scores: Option<&ScoresArtifact>,
-) -> SuggestionsArtifact {
-    let fallback_summary;
-    let fallback_metrics;
-    let (score_summary, score_metrics, score_version) = if let Some(scores) = scores {
-        (
-            &scores.summary,
-            &scores.metrics,
-            Some(scores.scores_version),
-        )
-    } else {
-        let failed_steps = run_summary.counts.failed_steps;
-        let success_steps = run_summary
-            .counts
-            .completed_steps
-            .saturating_sub(failed_steps);
-        let success_ratio = if run_summary.counts.total_steps == 0 {
-            1.0
-        } else {
-            let permille = (success_steps * 1000) / run_summary.counts.total_steps;
-            (permille as f64) / 1000.0
-        };
-        let security_denied_count: usize =
-            run_summary.policy.security_denials_by_code.values().sum();
-        let delegation_denied_count: usize = run_summary
-            .policy
-            .security_denials_by_code
-            .iter()
-            .filter_map(|(code, count)| {
-                if code.starts_with("DELEGATION_") {
-                    Some(*count)
-                } else {
-                    None
-                }
-            })
-            .sum();
-        fallback_summary = ScoresSummary {
-            success_ratio,
-            failure_count: failed_steps,
-            retry_count: 0,
-            delegation_denied_count,
-            security_denied_count,
-        };
-        fallback_metrics = ScoresMetrics {
-            scheduler_max_parallel_observed: 1,
-        };
-        (&fallback_summary, &fallback_metrics, None)
-    };
-
-    let base_evidence = SuggestionEvidence {
-        failure_count: score_summary.failure_count,
-        retry_count: score_summary.retry_count,
-        delegation_denied_count: score_summary.delegation_denied_count,
-        security_denied_count: score_summary.security_denied_count,
-        success_ratio: score_summary.success_ratio,
-        scheduler_max_parallel_observed: score_metrics.scheduler_max_parallel_observed,
-    };
-
-    let mut suggestions = Vec::new();
-
-    if score_summary.failure_count > 0 {
-        suggestions.push(SuggestionItem {
-            id: String::new(),
-            category: "retry".to_string(),
-            severity: "improvement".to_string(),
-            rationale: "One or more steps failed; consider safer retry policy for transient paths."
-                .to_string(),
-            evidence: base_evidence.clone(),
-            proposed_change: SuggestedChangeIntent {
-                intent: "increase_step_retry_budget".to_string(),
-                target: "failed-step-set".to_string(),
-            },
-        });
-    }
-    if score_summary.delegation_denied_count > 0 {
-        suggestions.push(SuggestionItem {
-            id: String::new(),
-            category: "delegation".to_string(),
-            severity: "warning".to_string(),
-            rationale: "Delegation-denied signals detected; review delegation policy scope."
-                .to_string(),
-            evidence: base_evidence.clone(),
-            proposed_change: SuggestedChangeIntent {
-                intent: "review_delegation_policy_scope".to_string(),
-                target: "delegation-boundary".to_string(),
-            },
-        });
-    }
-    if score_summary.security_denied_count > 0 {
-        suggestions.push(SuggestionItem {
-            id: String::new(),
-            category: "security".to_string(),
-            severity: "warning".to_string(),
-            rationale: "Security denials observed; align expected capabilities with trust policy."
-                .to_string(),
-            evidence: base_evidence.clone(),
-            proposed_change: SuggestedChangeIntent {
-                intent: "review_security_policy_expectations".to_string(),
-                target: "security-envelope".to_string(),
-            },
-        });
-    }
-    if score_summary.success_ratio < 1.0 {
-        suggestions.push(SuggestionItem {
-            id: String::new(),
-            category: "general".to_string(),
-            severity: "improvement".to_string(),
-            rationale: "Success ratio is below 1.0; review failing steps and dependency shape."
-                .to_string(),
-            evidence: base_evidence.clone(),
-            proposed_change: SuggestedChangeIntent {
-                intent: "review_failure_hotspots".to_string(),
-                target: "workflow-step-dependencies".to_string(),
-            },
-        });
-    }
-    if run_summary.counts.total_steps > 1 && score_metrics.scheduler_max_parallel_observed <= 1 {
-        suggestions.push(SuggestionItem {
-            id: String::new(),
-            category: "scheduler".to_string(),
-            severity: "info".to_string(),
-            rationale: "Observed parallelism is low; evaluate opportunities for safe concurrency."
-                .to_string(),
-            evidence: base_evidence,
-            proposed_change: SuggestedChangeIntent {
-                intent: "evaluate_parallelizable_dependencies".to_string(),
-                target: "workflow-structure".to_string(),
-            },
-        });
-    }
-
-    for (idx, suggestion) in suggestions.iter_mut().enumerate() {
-        suggestion.id = format!("sug-{:03}", idx + 1);
-    }
-
-    SuggestionsArtifact {
-        suggestions_version: SUGGESTIONS_VERSION,
-        run_id: run_summary.run_id.clone(),
-        generated_from: SuggestionsGeneratedFrom {
-            artifact_model_version: run_summary.artifact_model_version,
-            run_summary_version: run_summary.run_summary_version,
-            scores_version: score_version,
-        },
-        suggestions,
-    }
-}
+use super::*;
 
 fn aee_decision_kind_for_intent(intent: &str) -> (&'static str, &'static str) {
     match intent {
@@ -1906,689 +1284,4 @@ pub(crate) fn build_reasoning_graph_artifact(
             edges,
         },
     }
-}
-
-#[allow(clippy::too_many_arguments)]
-pub(crate) fn write_run_state_artifacts(
-    resolved: &resolve::AdlResolved,
-    tr: &trace::Trace,
-    adl_path: &Path,
-    _out_dir: &Path,
-    start_ms: u128,
-    end_ms: u128,
-    status: &str,
-    pause: Option<&execute::PauseState>,
-    steering_history: &[execute::SteeringRecord],
-    runtime_control: &execute::RuntimeControlState,
-    resume_completed_step_ids: Option<&HashSet<String>>,
-    failure: Option<&anyhow::Error>,
-) -> Result<PathBuf> {
-    let run_paths = artifacts::RunArtifactPaths::for_run(&resolved.run_id)?;
-    run_paths.ensure_layout()?;
-    run_paths.write_model_marker()?;
-    let run_dir = run_paths.run_dir();
-    let resume_completed: BTreeSet<String> = resume_completed_step_ids
-        .map(|ids| ids.iter().cloned().collect())
-        .unwrap_or_default();
-
-    let mut status_by_step: HashMap<String, String> = HashMap::new();
-    for ev in &tr.events {
-        if let trace::TraceEvent::StepFinished {
-            step_id, success, ..
-        } = ev
-        {
-            let status = if *success { "success" } else { "failure" };
-            status_by_step.insert(step_id.clone(), status.to_string());
-        }
-    }
-
-    let mut steps = Vec::with_capacity(resolved.steps.len());
-    for step in &resolved.steps {
-        let status = status_by_step
-            .get(&step.id)
-            .cloned()
-            .or_else(|| {
-                resume_completed
-                    .contains(&step.id)
-                    .then(|| "success".to_string())
-            })
-            .unwrap_or_else(|| "not_run".to_string());
-        let output_artifact_path = match (status.as_str(), step.write_to.as_deref()) {
-            ("success", Some(write_to)) => Some(write_to.to_string()),
-            _ => None,
-        };
-
-        let agent_id = step
-            .agent
-            .as_deref()
-            .unwrap_or("<unresolved-agent>")
-            .to_string();
-        let provider_id = step
-            .provider
-            .as_deref()
-            .unwrap_or("<unresolved-provider>")
-            .to_string();
-
-        steps.push(StepStateArtifact {
-            step_id: step.id.clone(),
-            agent_id,
-            provider_id,
-            status,
-            output_artifact_path,
-        });
-    }
-
-    let scheduler_policy = execute::scheduler_policy_for_run(resolved)?;
-    let error_message = tr.events.iter().rev().find_map(|ev| match ev {
-        trace::TraceEvent::RunFailed { message, .. } => Some(message.clone()),
-        _ => None,
-    });
-    let run_artifact = RunStateArtifact {
-        schema_version: RUN_STATE_SCHEMA_VERSION.to_string(),
-        run_id: resolved.run_id.clone(),
-        workflow_id: resolved.workflow_id.clone(),
-        version: resolved.doc.version.clone(),
-        status: status.to_string(),
-        error_message: error_message.clone(),
-        start_time_ms: start_ms,
-        end_time_ms: end_ms,
-        duration_ms: end_ms.saturating_sub(start_ms),
-        execution_plan_hash: execution_plan_hash(&resolved.execution_plan)?,
-        scheduler_max_concurrency: scheduler_policy.map(|(v, _)| v),
-        scheduler_policy_source: scheduler_policy.map(|(_, source)| source.as_str().to_string()),
-        steering_history: steering_history.to_vec(),
-        pause: pause.cloned(),
-    };
-
-    let run_json = serde_json::to_vec_pretty(&run_artifact).context("serialize run.json")?;
-    let steps_json = serde_json::to_vec_pretty(&steps).context("serialize steps.json")?;
-    let activation_log_path = run_paths.activation_log_json();
-    instrumentation::write_trace_artifact(&activation_log_path, &tr.events)?;
-    let cluster_groundwork = build_cluster_groundwork_artifact(resolved, &steps, tr);
-    let cluster_groundwork_json = serde_json::to_vec_pretty(&cluster_groundwork)
-        .context("serialize cluster_groundwork.json")?;
-    artifacts::atomic_write(
-        &run_paths.cluster_groundwork_json(),
-        &cluster_groundwork_json,
-    )?;
-    let run_summary = build_run_summary(
-        resolved,
-        status,
-        pause,
-        &steps,
-        tr.events
-            .iter()
-            .filter(|ev| matches!(ev, trace::TraceEvent::StepFinished { .. }))
-            .count(),
-        failure,
-        &run_paths,
-    );
-    let overall_status = match status {
-        "success" => "succeeded",
-        "failure" => "failed",
-        "paused" => "running",
-        other => other,
-    };
-    let run_status = build_run_status(
-        resolved,
-        tr,
-        overall_status,
-        &steps,
-        failure,
-        &resume_completed,
-    );
-    let run_summary_json =
-        serde_json::to_vec_pretty(&run_summary).context("serialize run_summary.json")?;
-    let run_status_json =
-        serde_json::to_vec_pretty(&run_status).context("serialize run_status.json")?;
-    let scores = build_scores_artifact(&run_summary, tr);
-    let scores_json = serde_json::to_vec_pretty(&scores).context("serialize scores.json")?;
-    let scores_for_suggestions = read_scores_if_present(&run_paths).unwrap_or(scores.clone());
-    let suggestions = build_suggestions_artifact(&run_summary, Some(&scores_for_suggestions));
-    let suggestions_json =
-        serde_json::to_vec_pretty(&suggestions).context("serialize suggestions.json")?;
-    let cognitive_signals = build_cognitive_signals_artifact_from_state(
-        &run_summary,
-        &runtime_control.signals,
-        &suggestions,
-        Some(&scores_for_suggestions),
-    );
-    let cognitive_signals_json = serde_json::to_vec_pretty(&cognitive_signals)
-        .context("serialize cognitive_signals.v1.json")?;
-    let affect_state =
-        build_affect_state_artifact(&run_summary, &suggestions, Some(&scores_for_suggestions));
-    let affect_state_json =
-        serde_json::to_vec_pretty(&affect_state).context("serialize affect_state.v1.json")?;
-    let cognitive_arbitration = build_cognitive_arbitration_artifact_from_state(
-        &run_summary,
-        &suggestions,
-        &runtime_control.arbitration,
-        Some(&scores_for_suggestions),
-    );
-    let fast_slow_path = build_fast_slow_path_artifact(
-        &run_summary,
-        &cognitive_arbitration,
-        &runtime_control.fast_slow,
-        Some(&scores_for_suggestions),
-    );
-    let agency_selection = build_agency_selection_artifact(
-        &run_summary,
-        &cognitive_arbitration,
-        &runtime_control.agency,
-        Some(&scores_for_suggestions),
-    );
-    let bounded_execution = build_bounded_execution_artifact(
-        &run_summary,
-        &fast_slow_path,
-        &agency_selection,
-        &runtime_control.bounded_execution,
-        Some(&scores_for_suggestions),
-    );
-    let evaluation_signals = build_evaluation_signals_artifact(
-        &run_summary,
-        &fast_slow_path,
-        &agency_selection,
-        &runtime_control.evaluation,
-        Some(&scores_for_suggestions),
-    );
-    let reframing = build_reframing_artifact(
-        &run_summary,
-        &fast_slow_path,
-        &agency_selection,
-        &runtime_control.reframing,
-        Some(&scores_for_suggestions),
-    );
-    let freedom_gate = build_freedom_gate_artifact(
-        &run_summary,
-        &evaluation_signals,
-        &runtime_control.freedom_gate,
-        Some(&scores_for_suggestions),
-    );
-    let memory_read = build_memory_read_artifact(
-        &run_summary,
-        &evaluation_signals,
-        &runtime_control.memory.read,
-        Some(&scores_for_suggestions),
-    );
-    let memory_write = build_memory_write_artifact(
-        &run_summary,
-        &evaluation_signals,
-        &runtime_control.memory.write,
-        Some(&scores_for_suggestions),
-    );
-    let control_path_memory =
-        build_control_path_memory_artifact(&run_summary, &memory_read, &memory_write);
-    let control_path_final_result = build_control_path_final_result_artifact(
-        &run_summary,
-        &cognitive_arbitration,
-        &agency_selection,
-        &evaluation_signals,
-        &freedom_gate,
-    );
-    let control_path_summary = build_control_path_summary(&ControlPathSummaryContext {
-        signals: &cognitive_signals,
-        agency: &agency_selection,
-        arbitration: &cognitive_arbitration,
-        execution: &bounded_execution,
-        evaluation: &evaluation_signals,
-        reframing: &reframing,
-        memory: &control_path_memory,
-        freedom_gate: &freedom_gate,
-        final_result: &control_path_final_result,
-    });
-    let cognitive_arbitration_json = serde_json::to_vec_pretty(&cognitive_arbitration)
-        .context("serialize cognitive_arbitration.v1.json")?;
-    let fast_slow_path_json =
-        serde_json::to_vec_pretty(&fast_slow_path).context("serialize fast_slow_path.v1.json")?;
-    let agency_selection_json = serde_json::to_vec_pretty(&agency_selection)
-        .context("serialize agency_selection.v1.json")?;
-    let bounded_execution_json = serde_json::to_vec_pretty(&bounded_execution)
-        .context("serialize bounded_execution.v1.json")?;
-    let evaluation_signals_json = serde_json::to_vec_pretty(&evaluation_signals)
-        .context("serialize evaluation_signals.v1.json")?;
-    let reframing_json =
-        serde_json::to_vec_pretty(&reframing).context("serialize reframing.v1.json")?;
-    let freedom_gate_json =
-        serde_json::to_vec_pretty(&freedom_gate).context("serialize freedom_gate.v1.json")?;
-    let memory_read_json =
-        serde_json::to_vec_pretty(&memory_read).context("serialize memory_read.v1.json")?;
-    let memory_write_json =
-        serde_json::to_vec_pretty(&memory_write).context("serialize memory_write.v1.json")?;
-    let control_path_memory_json = serde_json::to_vec_pretty(&control_path_memory)
-        .context("serialize control_path memory.json")?;
-    let control_path_final_result_json = serde_json::to_vec_pretty(&control_path_final_result)
-        .context("serialize control_path final_result.json")?;
-    let aee_decision = build_aee_decision_artifact(
-        &run_summary,
-        &suggestions,
-        &affect_state,
-        Some(&scores_for_suggestions),
-    );
-    let reasoning_graph = build_reasoning_graph_artifact(
-        &run_summary,
-        &affect_state,
-        &aee_decision,
-        Some(&scores_for_suggestions),
-    );
-    let aee_decision_json =
-        serde_json::to_vec_pretty(&aee_decision).context("serialize aee_decision.json")?;
-    let reasoning_graph_json =
-        serde_json::to_vec_pretty(&reasoning_graph).context("serialize reasoning_graph.v1.json")?;
-
-    artifacts::atomic_write(&run_paths.run_json(), &run_json)?;
-    artifacts::atomic_write(&run_paths.steps_json(), &steps_json)?;
-    artifacts::atomic_write(&run_paths.run_status_json(), &run_status_json)?;
-    artifacts::atomic_write(&run_paths.run_summary_json(), &run_summary_json)?;
-    artifacts::atomic_write(&run_paths.scores_json(), &scores_json)?;
-    artifacts::atomic_write(&run_paths.suggestions_json(), &suggestions_json)?;
-    artifacts::atomic_write(&run_paths.cognitive_signals_json(), &cognitive_signals_json)?;
-    artifacts::atomic_write(
-        &run_paths.cognitive_arbitration_json(),
-        &cognitive_arbitration_json,
-    )?;
-    artifacts::atomic_write(&run_paths.fast_slow_path_json(), &fast_slow_path_json)?;
-    artifacts::atomic_write(&run_paths.agency_selection_json(), &agency_selection_json)?;
-    artifacts::atomic_write(&run_paths.bounded_execution_json(), &bounded_execution_json)?;
-    artifacts::atomic_write(
-        &run_paths.evaluation_signals_json(),
-        &evaluation_signals_json,
-    )?;
-    artifacts::atomic_write(&run_paths.reframing_json(), &reframing_json)?;
-    artifacts::atomic_write(&run_paths.freedom_gate_json(), &freedom_gate_json)?;
-    artifacts::atomic_write(&run_paths.memory_read_json(), &memory_read_json)?;
-    artifacts::atomic_write(&run_paths.memory_write_json(), &memory_write_json)?;
-    artifacts::atomic_write(
-        &run_paths.control_path_signals_json(),
-        &cognitive_signals_json,
-    )?;
-    artifacts::atomic_write(
-        &run_paths.control_path_candidate_selection_json(),
-        &agency_selection_json,
-    )?;
-    artifacts::atomic_write(
-        &run_paths.control_path_arbitration_json(),
-        &cognitive_arbitration_json,
-    )?;
-    artifacts::atomic_write(
-        &run_paths.control_path_execution_iterations_json(),
-        &bounded_execution_json,
-    )?;
-    artifacts::atomic_write(
-        &run_paths.control_path_evaluation_json(),
-        &evaluation_signals_json,
-    )?;
-    artifacts::atomic_write(&run_paths.control_path_reframing_json(), &reframing_json)?;
-    artifacts::atomic_write(
-        &run_paths.control_path_memory_json(),
-        &control_path_memory_json,
-    )?;
-    artifacts::atomic_write(
-        &run_paths.control_path_freedom_gate_json(),
-        &freedom_gate_json,
-    )?;
-    artifacts::atomic_write(
-        &run_paths.control_path_final_result_json(),
-        &control_path_final_result_json,
-    )?;
-    artifacts::atomic_write(
-        &run_paths.control_path_summary_txt(),
-        control_path_summary.as_bytes(),
-    )?;
-    artifacts::atomic_write(&run_paths.cognitive_signals_json(), &cognitive_signals_json)?;
-    artifacts::atomic_write(
-        &run_paths.cognitive_arbitration_json(),
-        &cognitive_arbitration_json,
-    )?;
-    artifacts::atomic_write(&run_paths.fast_slow_path_json(), &fast_slow_path_json)?;
-    artifacts::atomic_write(&run_paths.agency_selection_json(), &agency_selection_json)?;
-    artifacts::atomic_write(&run_paths.bounded_execution_json(), &bounded_execution_json)?;
-    artifacts::atomic_write(
-        &run_paths.evaluation_signals_json(),
-        &evaluation_signals_json,
-    )?;
-    artifacts::atomic_write(&run_paths.reframing_json(), &reframing_json)?;
-    artifacts::atomic_write(&run_paths.freedom_gate_json(), &freedom_gate_json)?;
-    artifacts::atomic_write(&run_paths.memory_read_json(), &memory_read_json)?;
-    artifacts::atomic_write(&run_paths.memory_write_json(), &memory_write_json)?;
-    artifacts::atomic_write(&run_paths.affect_state_json(), &affect_state_json)?;
-    artifacts::atomic_write(&run_paths.aee_decision_json(), &aee_decision_json)?;
-    artifacts::atomic_write(&run_paths.reasoning_graph_json(), &reasoning_graph_json)?;
-    if let Some(pause_payload) = pause {
-        let pause_artifact = PauseStateArtifact {
-            schema_version: PAUSE_STATE_SCHEMA_VERSION.to_string(),
-            run_id: resolved.run_id.clone(),
-            workflow_id: resolved.workflow_id.clone(),
-            version: resolved.doc.version.clone(),
-            status: "paused".to_string(),
-            adl_path: sanitize_pause_adl_path(adl_path),
-            execution_plan_hash: execution_plan_hash(&resolved.execution_plan)?,
-            steering_history: steering_history.to_vec(),
-            pause: pause_payload.clone(),
-        };
-        let pause_json =
-            serde_json::to_vec_pretty(&pause_artifact).context("serialize pause_state.json")?;
-        artifacts::atomic_write(&run_paths.pause_state_json(), &pause_json)?;
-    }
-
-    Ok(run_dir)
-}
-
-fn read_required_json_artifact<T>(control_path_dir: &Path, file_name: &str) -> Result<T>
-where
-    T: for<'de> Deserialize<'de>,
-{
-    let path = control_path_dir.join(file_name);
-    let raw = std::fs::read_to_string(&path).with_context(|| {
-        format!(
-            "missing required control-path artifact '{}'",
-            path.display()
-        )
-    })?;
-    serde_json::from_str(&raw)
-        .with_context(|| format!("invalid control-path artifact '{}'", path.display()))
-}
-
-pub(crate) fn validate_control_path_artifact_set(control_path_dir: &Path) -> Result<()> {
-    if !control_path_dir.exists() {
-        return Err(anyhow!(
-            "control-path artifact root does not exist: {}",
-            control_path_dir.display()
-        ));
-    }
-
-    let signals: CognitiveSignalsArtifact =
-        read_required_json_artifact(control_path_dir, "signals.json")?;
-    let agency: AgencySelectionArtifact =
-        read_required_json_artifact(control_path_dir, "candidate_selection.json")?;
-    let arbitration: CognitiveArbitrationArtifact =
-        read_required_json_artifact(control_path_dir, "arbitration.json")?;
-    let execution: BoundedExecutionArtifact =
-        read_required_json_artifact(control_path_dir, "execution_iterations.json")?;
-    let evaluation: EvaluationSignalsArtifact =
-        read_required_json_artifact(control_path_dir, "evaluation.json")?;
-    let reframing: ReframingArtifact =
-        read_required_json_artifact(control_path_dir, "reframing.json")?;
-    let memory: ControlPathMemoryArtifact =
-        read_required_json_artifact(control_path_dir, "memory.json")?;
-    let freedom_gate: FreedomGateArtifact =
-        read_required_json_artifact(control_path_dir, "freedom_gate.json")?;
-    let final_result: ControlPathFinalResultArtifact =
-        read_required_json_artifact(control_path_dir, "final_result.json")?;
-
-    let summary_path = control_path_dir.join("summary.txt");
-    let summary = std::fs::read_to_string(&summary_path).with_context(|| {
-        format!(
-            "missing required control-path artifact '{}'",
-            summary_path.display()
-        )
-    })?;
-    if summary.trim().is_empty() {
-        return Err(anyhow!(
-            "control-path summary is empty at '{}'",
-            summary_path.display()
-        ));
-    }
-
-    let expected_stage_order = vec![
-        "signals".to_string(),
-        "candidate_selection".to_string(),
-        "arbitration".to_string(),
-        "execution".to_string(),
-        "evaluation".to_string(),
-        "reframing".to_string(),
-        "memory".to_string(),
-        "freedom_gate".to_string(),
-        "final_result".to_string(),
-    ];
-    if final_result.stage_order != expected_stage_order {
-        return Err(anyhow!(
-            "control-path final_result stage_order mismatch: expected {:?}, found {:?}",
-            expected_stage_order,
-            final_result.stage_order
-        ));
-    }
-
-    let run_ids = [
-        signals.run_id.as_str(),
-        agency.run_id.as_str(),
-        arbitration.run_id.as_str(),
-        execution.run_id.as_str(),
-        evaluation.run_id.as_str(),
-        reframing.run_id.as_str(),
-        memory.run_id.as_str(),
-        freedom_gate.run_id.as_str(),
-        final_result.run_id.as_str(),
-    ];
-    let canonical_run_id = final_result.run_id.as_str();
-    if run_ids.iter().any(|run_id| *run_id != canonical_run_id) {
-        return Err(anyhow!(
-            "control-path artifact run_id mismatch under '{}'",
-            control_path_dir.display()
-        ));
-    }
-
-    if final_result.route_selected != arbitration.route_selected {
-        return Err(anyhow!(
-            "control-path final_result route '{}' does not match arbitration route '{}'",
-            final_result.route_selected,
-            arbitration.route_selected
-        ));
-    }
-    if final_result.selected_candidate != agency.selected_candidate_id {
-        return Err(anyhow!(
-            "control-path final_result selected_candidate '{}' does not match candidate_selection '{}'",
-            final_result.selected_candidate,
-            agency.selected_candidate_id
-        ));
-    }
-    if final_result.termination_reason != evaluation.termination_reason {
-        return Err(anyhow!(
-            "control-path final_result termination_reason '{}' does not match evaluation '{}'",
-            final_result.termination_reason,
-            evaluation.termination_reason
-        ));
-    }
-    if final_result.gate_decision != freedom_gate.gate_decision {
-        return Err(anyhow!(
-            "control-path final_result gate_decision '{}' does not match freedom_gate '{}'",
-            final_result.gate_decision,
-            freedom_gate.gate_decision
-        ));
-    }
-    if final_result.next_control_action != evaluation.next_control_action {
-        return Err(anyhow!(
-            "control-path final_result next_control_action '{}' does not match evaluation '{}'",
-            final_result.next_control_action,
-            evaluation.next_control_action
-        ));
-    }
-
-    let required_summary_markers = [
-        "stage_order: signals -> candidate_selection -> arbitration -> execution -> evaluation -> reframing -> memory -> freedom_gate -> final_result".to_string(),
-        format!("candidate_selection: candidate_id={}", agency.selected_candidate_id),
-        format!("arbitration: route={}", arbitration.route_selected),
-        format!("evaluation: termination_reason={}", evaluation.termination_reason),
-        format!("reframing: trigger={}", reframing.reframing_trigger),
-        format!("freedom_gate: decision={}", freedom_gate.gate_decision),
-        format!("final_result: {}", final_result.final_result),
-    ];
-    for marker in required_summary_markers {
-        if !summary.contains(&marker) {
-            return Err(anyhow!(
-                "control-path summary '{}' is missing required marker '{}'",
-                summary_path.display(),
-                marker
-            ));
-        }
-    }
-
-    Ok(())
-}
-
-pub(crate) fn load_resume_state(
-    path: &Path,
-    resolved: &resolve::AdlResolved,
-) -> Result<execute::ResumeState> {
-    let raw = std::fs::read_to_string(path)
-        .with_context(|| format!("failed to read resume state '{}'", path.display()))?;
-    let artifact: RunStateArtifact = serde_json::from_str(&raw).with_context(|| {
-        format!(
-            "failed to parse resume state '{}' as run_state artifact",
-            path.display()
-        )
-    })?;
-
-    if artifact.schema_version != RUN_STATE_SCHEMA_VERSION {
-        return Err(anyhow::anyhow!(
-            "resume state schema_version mismatch: state='{}' expected='{}'",
-            artifact.schema_version,
-            RUN_STATE_SCHEMA_VERSION
-        ));
-    }
-
-    if artifact.schema_version != RUN_STATE_SCHEMA_VERSION {
-        return Err(anyhow::anyhow!(
-            "resume state schema_version mismatch: state='{}' expected='{}'",
-            artifact.schema_version,
-            RUN_STATE_SCHEMA_VERSION
-        ));
-    }
-
-    if artifact.status != "paused" {
-        return Err(anyhow::anyhow!(
-            "resume state must have status='paused' (found='{}' for run_id='{}' in '{}')",
-            artifact.status,
-            artifact.run_id,
-            path.display()
-        ));
-    }
-    if artifact.run_id != resolved.run_id {
-        return Err(anyhow::anyhow!(
-            "resume run_id mismatch: state='{}' current='{}'",
-            artifact.run_id,
-            resolved.run_id
-        ));
-    }
-    if artifact.workflow_id != resolved.workflow_id {
-        return Err(anyhow::anyhow!(
-            "resume workflow_id mismatch for run_id='{}' in '{}': state='{}' current='{}'",
-            artifact.run_id,
-            path.display(),
-            artifact.workflow_id,
-            resolved.workflow_id
-        ));
-    }
-    if artifact.version != resolved.doc.version {
-        return Err(anyhow::anyhow!(
-            "resume version mismatch for run_id='{}' in '{}': state='{}' current='{}'",
-            artifact.run_id,
-            path.display(),
-            artifact.version,
-            resolved.doc.version
-        ));
-    }
-    let plan_hash = execution_plan_hash(&resolved.execution_plan)?;
-    if artifact.execution_plan_hash != plan_hash {
-        return Err(anyhow::anyhow!(
-            "resume execution plan mismatch for run_id='{}' in '{}'; state plan != current plan (resume requires identical plan + ordering)",
-            artifact.run_id,
-            path.display()
-        ));
-    }
-    let pause = artifact
-        .pause
-        .ok_or_else(|| anyhow::anyhow!("resume state missing pause payload"))?;
-
-    let completed_step_ids = pause.completed_step_ids.into_iter().collect();
-    Ok(execute::ResumeState {
-        completed_step_ids,
-        saved_state: pause.saved_state,
-        completed_outputs: pause.completed_outputs,
-        steering_history: artifact.steering_history,
-    })
-}
-
-pub(crate) fn resume_state_path_for_run_id(run_id: &str) -> Result<PathBuf> {
-    Ok(artifacts::RunArtifactPaths::for_run(run_id)?.pause_state_json())
-}
-
-pub(crate) fn load_pause_state_artifact(path: &Path) -> Result<PauseStateArtifact> {
-    let raw = std::fs::read_to_string(path)
-        .with_context(|| format!("failed to read pause state '{}'", path.display()))?;
-    let artifact: PauseStateArtifact =
-        serde_json::from_str(&raw).with_context(|| "failed to parse pause_state.json")?;
-    Ok(artifact)
-}
-
-pub(crate) fn load_steering_patch(path: &Path) -> Result<(execute::SteeringPatch, String)> {
-    let raw = std::fs::read(path)
-        .with_context(|| format!("failed to read steering patch '{}'", path.display()))?;
-    let fingerprint = stable_fingerprint_hex(&raw);
-    let patch: execute::SteeringPatch =
-        serde_json::from_slice(&raw).with_context(|| "failed to parse steering patch JSON")?;
-    execute::validate_steering_patch(&patch)?;
-    Ok((patch, fingerprint))
-}
-
-pub(crate) fn validate_pause_artifact_basic(
-    artifact: &PauseStateArtifact,
-    run_id: &str,
-) -> Result<()> {
-    if artifact.schema_version != PAUSE_STATE_SCHEMA_VERSION {
-        return Err(anyhow::anyhow!(
-            "pause state schema_version mismatch: state='{}' expected='{}'",
-            artifact.schema_version,
-            PAUSE_STATE_SCHEMA_VERSION
-        ));
-    }
-    if artifact.status != "paused" {
-        return Err(anyhow::anyhow!(
-            "pause state must have status='paused' (found '{}')",
-            artifact.status
-        ));
-    }
-    if artifact.run_id != run_id {
-        return Err(anyhow::anyhow!(
-            "pause state run_id mismatch: state='{}' requested='{}'",
-            artifact.run_id,
-            run_id
-        ));
-    }
-    Ok(())
-}
-
-pub(crate) fn validate_pause_artifact_for_resume(
-    artifact: &PauseStateArtifact,
-    run_id: &str,
-    resolved: &resolve::AdlResolved,
-) -> Result<()> {
-    validate_pause_artifact_basic(artifact, run_id)?;
-    if artifact.run_id != resolved.run_id {
-        return Err(anyhow::anyhow!(
-            "resume run_id mismatch: state='{}' current='{}'",
-            artifact.run_id,
-            resolved.run_id
-        ));
-    }
-    if artifact.workflow_id != resolved.workflow_id {
-        return Err(anyhow::anyhow!(
-            "resume workflow_id mismatch: state='{}' current='{}'",
-            artifact.workflow_id,
-            resolved.workflow_id
-        ));
-    }
-    if artifact.version != resolved.doc.version {
-        return Err(anyhow::anyhow!(
-            "resume version mismatch: state='{}' current='{}'",
-            artifact.version,
-            resolved.doc.version
-        ));
-    }
-    let plan_hash = execution_plan_hash(&resolved.execution_plan)?;
-    if artifact.execution_plan_hash != plan_hash {
-        return Err(anyhow::anyhow!(
-            "resume execution plan hash mismatch; resume requires identical plan and ordering"
-        ));
-    }
-    Ok(())
 }
